Remove unsigned ints

+++ src/engine/card.hpp
@@ -19,17 +19,17 @@ enum Faction {
 class Card {
     private:
         string name_;
-        uint16_t power_;
+        int16_t power_;
         Range range_;
         string* ability_;
         bool is_special_;
         Faction faction_;
     public:
         Card();
-        Card(string name, uint8_t power, Range range, string* ability, bool is_special, Faction faction);
+        Card(string name, int16_t power, Range range, string* ability, bool is_special, Faction faction);
         void trigger_ability();
         [[nodiscard]] Range getRange() const;
-        [[nodiscard]] uint16_t getPower() const;
+        [[nodiscard]] int16_t getPower() const;
         [[nodiscard]] string getName() const;
         [[nodiscard]] bool isSpecial() const;
         [[nodiscard]] Faction getFaction() const;
